+++ platform/windows/native.h
@@ -7,6 +7,18 @@ extern "C" {
 
 typedef struct VitraWin VitraWin;
 
+typedef struct {
+	char *title;
+	int width;
+	int height;
+	int maximized;
+	int fullscreen;
+	int above;
+	int minimized;
+	int hidden;
+	char *icon_path;
+} VitraChrome;
+
 void vitra_win32_init(void);
 void vitra_win32_main(void);
 void vitra_win32_quit(void);
@@ -17,6 +29,8 @@ void vitra_win_navigate(VitraWin *w, const char *uri);
 int vitra_win_eval(VitraWin *w, const char *js);
 void vitra_win_close(VitraWin *w);
 void vitra_win_free(VitraWin *w);
+void vitra_win_apply_chrome(VitraWin *w, const char *title, int width, int height, int maximized, int fullscreen, int above, int minimized, int hidden, const char *icon_path);
+VitraChrome vitra_win_chrome(VitraWin *w);
 
 #ifdef __cplusplus
 }